Split the functions out into raw.py.
